refactor(firmware-vue): per-component folder layout
Move each component into its own camelCase folder named after the
component minus the Astros prefix, matching the existing pattern for
AstrosLayout (common/layout/), AstrosSearchSelect (common/searchableSelect/),
and the field components. The .vue, .stories.ts, and component-scoped
types.ts live together inside the folder; the firmware/index.ts barrel
re-exports through the deeper paths.

  firmware/
  ├── index.ts
  ├── firmwareButton/
  │   ├── AstrosFirmwareButton.vue
  │   ├── AstrosFirmwareButton.stories.ts
  │   └── types.ts
  └── firmwareSourceStrip/
      ├── AstrosFirmwareSourceStrip.vue
      └── AstrosFirmwareSourceStrip.stories.ts

All 5 file moves detected as renames (R) by git so history follows.
Pure structural change; no behavior modification. Build, tests, format,
lint all green.
