+++ astros_vue/src/components/firmware/firmwareButton/AstrosFirmwareButton.vue
