+++ astros_vue/src/components/firmware/firmwareButton/types.ts
