+++ astros_vue/src/components/firmware/firmwareSourceStrip/AstrosFirmwareSourceStrip.stories.ts
