+++ astros_vue/src/components/firmware/index.ts
@@ -1,3 +1,3 @@
-export { default as AstrosFirmwareButton } from './AstrosFirmwareButton.vue';
-export { default as AstrosFirmwareSourceStrip } from './AstrosFirmwareSourceStrip.vue';
-export type { AstrosFirmwareButtonKind } from './types';
+export { default as AstrosFirmwareButton } from './firmwareButton/AstrosFirmwareButton.vue';
+export { default as AstrosFirmwareSourceStrip } from './firmwareSourceStrip/AstrosFirmwareSourceStrip.vue';
+export type { AstrosFirmwareButtonKind } from './firmwareButton/types';
